Move `applicationContext` & mirror it for testing
This moves `applicationContext` to the `utils` package.

Additionally, it is mirrored in the tests to allow late initialization.
This will be used in the following commit.

This is highly experimental.
I have no idea what I am doing.
