+++ app/src/main/java/com/ubergeek42/WeechatAndroid/utils/ApplicationContext.kt
@@ -1,0 +1,6 @@
+package com.ubergeek42.WeechatAndroid.utils
+
+import android.content.Context
+import com.ubergeek42.WeechatAndroid.Weechat
+
+val applicationContext: Context = Weechat.applicationContext
