+++ app/src/main/java/com/ubergeek42/WeechatAndroid/utils/MigratePreferences.kt
@@ -5,7 +5,6 @@ import androidx.core.content.edit
 import androidx.preference.PreferenceManager
 import androidx.preference.PrivateKeyPickerPreference
 import com.ubergeek42.WeechatAndroid.service.P.VolumeRole
-import com.ubergeek42.WeechatAndroid.upload.applicationContext
 import com.ubergeek42.WeechatAndroid.utils.AndroidKeyStoreUtils.InsideSecureHardware
 import com.ubergeek42.cats.Kitty
 import com.ubergeek42.cats.Root
