+++ app/src/main/java/com/ubergeek42/WeechatAndroid/views/CircularImageButton.kt
@@ -10,7 +10,7 @@ import android.view.ViewOutlineProvider
 import android.view.animation.Animation
 import android.view.animation.ScaleAnimation
 import androidx.appcompat.widget.AppCompatImageButton
-import com.ubergeek42.WeechatAndroid.upload.applicationContext
+import com.ubergeek42.WeechatAndroid.utils.applicationContext
 import com.ubergeek42.WeechatAndroid.upload.f
