+++ app/src/testDebug/java/com/ubergeek42/WeechatAndroid/utils/ApplicationContext.kt
@@ -1,0 +1,5 @@
+package com.ubergeek42.WeechatAndroid.utils
+
+import android.content.Context
+
+lateinit var applicationContext: Context
